feat: Get one file
