+++ src/frformat/geo/code_commune_insee.py
@@ -1,9 +1,9 @@
-from frformat import geo_format
+from frformat import new_format
 from frformat.geo.code_commune_insee_frozenset import (
     CODES_COMMUNES_INSEE_COG_2023,
     CODES_COMMUNES_INSEE_COG_2024,
 )
-from frformat.geo_format import Millesime
+from frformat.new_format import Millesime
 from frformat.versioned_set import VersionedSet
 
 name = "Code commune INSEE"
@@ -17,6 +17,6 @@ code_commune_insee_versioned_data.add_version(
     Millesime.M2024, CODES_COMMUNES_INSEE_COG_2024
 )
 
-CodeCommuneInsee = geo_format.new(
+CodeCommuneInsee = new_format.new(
     "CodeCommuneInsee", name, description, code_commune_insee_versioned_data
 )
