+++ src/frformat/geo/commune.py
@@ -1,6 +1,6 @@
-from frformat import geo_format
+from frformat import new_format
 from frformat.geo.commune_frozenset import COMMUNES_COG_2023, COMMUNES_COG_2024
-from frformat.geo_format import Millesime
+from frformat.new_format import Millesime
 from frformat.versioned_set import VersionedSet
 
 name = "Nom de commune"
@@ -13,4 +13,4 @@ commune_versioned_data = VersionedSet[Millesime]()
 commune_versioned_data.add_version(Millesime.M2023, COMMUNES_COG_2023)
 commune_versioned_data.add_version(Millesime.M2024, COMMUNES_COG_2024)
 
-Commune = geo_format.new("Commune", name, description, commune_versioned_data)
+Commune = new_format.new("Commune", name, description, commune_versioned_data)
