+++ src/frformat/geo/numero_departement.py
@@ -1,5 +1,5 @@
-from frformat import geo_format
-from frformat.geo_format import Millesime
+from frformat import new_format
+from frformat.new_format import Millesime
 from frformat.versioned_set import VersionedSet
 
 NUMEROS_DEPARTEMENTS_COG_2023 = frozenset(
@@ -24,6 +24,6 @@ numero_departement_versioned_data.add_version(
     Millesime.M2024, NUMEROS_DEPARTEMENTS_COG_2024
 )
 
-NumeroDepartement = geo_format.new(
+NumeroDepartement = new_format.new(
     "NumeroDepartement", name, description, numero_departement_versioned_data
 )
